requirements.txt
requests
kafka-python

kafka-python
polars
pandas
pandas-ta
duckdb
toolz
rx
